### 2025-09-29 #9097 @editorialbot
[PRE REVIEW]: PIVA: Photoemission Interface for Visualization and Analysis
**Submitting author:** @pudeIko (<a href="http://orcid.org/0000-0001-5889-4594">Wojciech Pudelko</a>)
**Repository:** https://github.com/pudeIko/piva
**Branch with paper.md** (empty if default branch): main
**Version:** v3.0.0
