### 2025-09-30 #9097 @pudeIko
Hi @kyleniemeyer ,
Thank you very much for your message and for taking the time to look into my submission!

I’ll go through the list of suggested reviewers and get back to you with my selections.
I have one question, possibly also related to the above: the software package has already been reviewed and accepted by pyOpenSci — how exactly does this affect the review process at JOSS?
