### 2025-10-01 #9097 @kyleniemeyer
This submission has already been reviewed and accepted by pyOpenSci (https://github.com/pyOpenSci/software-submission/issues/231) and is in scope for JOSS, so its publication here will be fast tracked.
