### 2025-10-01 #9129 @kyleniemeyer
@pudeIko could you fix the Numba citation in the paper? Some capitalizations in the title need `{}` to protect, and the DOI 10.1145/2833157.2833162 should be added
