### 2025-10-02 f50d7f8
makes changes in citations and adds badges in README
