+++ paper/paper.bib
@@ -92,10 +92,11 @@
 
 @inproceedings{numba,
   author        = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
-  title         = {Numba: A llvm-based python jit compiler},
+  title         = {Numba: a {LLVM}-based {P}ython {JIT} compiler},
   booktitle     = {Proceedings of the {S}econd {W}orkshop on the {LLVM} {C}ompiler {I}nfrastructure in {HPC}},
   pages         = {1--6},
-  year          = {2015}
+  year          = {2015},
+  doi           = {10.1145/2833157.2833162}
 }
 
 @article{numpy,
